+++ src/VaultSharp/V1/SecretsEngines/Active Directory/ActiveDirectorySecretsEngineProvider.cs
@@ -1,0 +1,22 @@
+﻿using System.Net.Http;
+using System.Threading.Tasks;
+using VaultSharp.Core;
+using VaultSharp.V1.Commons;
+
+namespace VaultSharp.V1.SecretsEngines.ActiveDirectory
+{
+    internal class ActiveDirectorySecretsEngineProvider : IActiveDirectorySecretsEngine
+    {
+        private readonly Polymath _polymath;
+
+        public ActiveDirectorySecretsEngineProvider(Polymath polymath)
+        {
+            _polymath = polymath;
+        }
+
+        public async Task<Secret<ActiveDirectoryCredentials>> GetCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.ActiveDirectory, string wrapTimeToLive = null)
+        {
+            return await _polymath.MakeVaultApiRequest<Secret<ActiveDirectoryCredentials>>("v1/" + mountPoint.Trim('/') + "/creds/" + roleName.Trim('/'), HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+    }
+}